x86: svm_cpu_park() does not write to guest memory
Reworked svm_cpu_park() according to the discussion followed commit
292746c on Github. To summarize:

 - There are no VMX guest activity states equivalent in AMD-V, thus
   no easy way to halt a CPU in guest mode waiting for NMI.

 - Copying parking code directly into guest memory is ugly and
   dangerous, and it is not to say Linux may see changes made by
   svm_cpu_park() as memory corruption.

 - There is no straightforward way to halt CPU waiting for NMI in host
   mode - events can be lost.

The approach implemented by this commit is probably the least evil: we
create a static parking code page and static shared page tables that maps it
at reset vector (0xfffffff0). When a CPU is parked, its NPT is switched to
that shared page tables, and CS:RIP are set so the code at reset vector is
executed. When the CPU resumes, NPTs are switched back to real ones.