x86: Prepare for link-time decisions on AMD-V/VMX
These changes preapre the x86 codebase for link-time decisions on
using AMD-V or VMX implementations for variant-dependent functions.
Since they are built on previous work, modifications to existing
upstream (VMX) code and current AMD-V code are made in a single commit.

NMI must be treated differently for AMD-V and SVM, so apic_nmi_handler()
was deleted and {vmx,svm}_nmi_handler() was introduced. On the contrary,
set_cs() can be made generic, so #ifdef can be removed safely.