configs: Reintroduce QEMU VM 2.0 config
My testing environment still uses QEMU 2.0, so I decided to have
a config separate from upstream to play with. It is in "works for me"
state and may have certain things broken, especially in VMX/VTD support
(the primary focus for me is AMD-V for obvious reasons).

qemu-vm-2.0 should be considered as a temporary measure until I move my
testing setup to 2.1, and is very likely to disappear before merge.