tools: config-create: Initial support for non-Intel boards
DMAR is Intel-specific. For AMD, we need to parse IVRS table, but
this code is currently missing since there is no IOMMU support in
AMD-V port.

Initially, we just make DMAR optional for AMD and modified the template
so missing bits of information are not included (and IOAPIC ID is set
to zero).

Signed-off-by: Valentine Sinitsyn <valentine.sinitsyn@gmail.com>