x86: Prevent unnecessary rebuilds in Makefile
Previously, both built-in-amd.o and built-in-intel.o where linked
from objects of the same names, and as each make invocation builds
both variants, this inevitably led to rebuild of the arch-specific
part, regardless if there were changes in the sources or not.

New Makefile use separate object files for Intel and AMD, and doesn't
break dependency tracking anymore. This way, only changed parts of
Jailhouse are recompiled and relinked at the each subsequent build.