configs: Make configs adjustable with config.h
As the number of cell configs grows, it becomes harder and harder
to make them generic enough to run on every supported hardware
configuration due to resource confilcts.

Attempt to workaroud this with introducing config.h to cell configs.
With this in place, some parts of cell configs (eg. I/O ranges) can
be compiled out on each particular system, removing the need for manual
changes in the configs.