x86: fix alias handling in AMD IOMMU code
AMD IOMMU permits aliases in IVRS tables. These are devices which
pretend to have device IDs different from their real ones when
communicating to the IOMMU.

Jailhouse provided support for aliases since the beginning, but as this
code was never really tested, it did it in a wrong way. In particular,
it didn't distinguish between aliased and real device IDs.

This patch introduces alias_bdf member to struct jailhouse_pci_device
and makes use of it both in the config tool and PCI support code.

The code is still compile-tested only, though.

Signed-off-by: Valentine Sinitsyn <valentine.sinitsyn@gmail.com>